refactor: remove deprecated builder
