+++ agents.log
@@ -1,3 +1,4 @@
+AGENT NOTE - 2025-07-12: Removed deprecated builder module and updated imports
 AGENT NOTE - 2025-07-16: Revised built-in resource config validation and tests
 AGENT NOTE - 2025-07-12: Added infrastructure deps injection in ResourceContainer
 AGENT NOTE - 2025-07-14: Unified builder and runtime inside Agent
